"restore deleted" action
